# src/main/resources/swagger/api.yml
openapi: 3.0.0
info:
  title: Client API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
    description: Local Server

paths:
  /clients:
    get:
      summary: Get a list of clients
      operationId: getClients
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientDTO'

    post:
      summary: Create a new client
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDTO'
      responses:
        '201':
          description: Client created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'

  /clients/{clientId}:
    get:
      summary: Get a specific client by ID
      operationId: getClientById
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'

    put:
      summary: Update a specific client by ID
      operationId: updateClientById
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDTO'
      responses:
        '200':
          description: Client updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'

    delete:
      summary: Delete a specific client by ID
      operationId: deleteClientById
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Client deleted successfully

components:
  schemas:
    ClientDTO:
      type: object
      properties:
        clientId:
          type: integer
          format: int64
        clientName:
          type: string
          pattern: "^(?!\\s*$)[^\\s]+$" # Client Name cannot contain spaces
        currency:
          type: string
        billingMethod:
          type: string
        emailId:
          type: string
          format: email
        firstName:
          type: string
          pattern: "^(?!\\s*$)[^\\s]+$" # First Name cannot contain spaces
        lastName:
          type: string
          pattern: "^(?!\\s*$)[^\\s]+$" # Last Name cannot contain spaces
        phone:
          type: string
        mobile:
          type: string
          pattern: "^[0-9]{10}$" # Mobile must be exactly 10 digits
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
        industry:
          type: string
        companySize:
          type: integer
        description:
          type: string